+++ test/smokes/swiftlint/expectations.rb
@@ -146,7 +146,7 @@ s.add_test("no_swift_file", type: "success", issues: [], analyzer: { name: "Swif
 s.add_test(
   "no_config_file",
   type: "failure",
-  message: "SwiftLint unexpectedly failed. Please see the log for details.",
+  message: "Could not read configuration file at path 'not_found.yml'.",
   analyzer: { name: "SwiftLint", version: "0.39.2" },
   warnings: [
     {
@@ -172,6 +172,6 @@ s.add_test(
 s.add_test(
   "wrong_swiftlint_version_set",
   type: "failure",
-  message: "SwiftLint unexpectedly failed. Please see the log for details.",
+  message: "Currently running SwiftLint 0.39.2 but configuration specified version 0.0.0.",
   analyzer: { name: "SwiftLint", version: "0.39.2" }
 )